Fixed a rare filter bug with sharpening kernels near EVs that could cause the result to spike up.
Details: In the original filter code (as described in the paper)
corner faces near EVs were ignored and only immediate neighbors were
filtered.  This created a slight discontinuity near EVs that was
generally invisible but became visible with large blurs.  More
recently (but before the public release), an improved method was
implemented that applied the corner portion of the kernel to all of
the corner faces but first processed the corner kernel to make it
symmetric.  This algorithm was robust with positive (non-sharpening)
filters like gaussian, but could introduce an artifact with sharpening
filters like mitchell.  When making the corner symmetric,
positive/negative weight pairs could cancel and produce near-zero
weight. When rescaling the kernel to preserve the original overall
weight, this could cause the weights to spike up arbitrarily high
producing spikes in the output.  The fix clamps the rescaling amount
to prevent the spikes.  Note: the clamping has no effect in the
normal, non-spike case, so there should be no visible effect due to
this change other than removal of the artifact.