Makefile: Create universal binaries on OS X
Adjust ptex/Makefile and utils/Makefile so that they include a
common etc/common.mak.

Detect OS X and adjust flags so that universal binaries are output.
Name the DSO "libPtex.dylib" on OS X and "libPtex.so" elsewhere.
Use libtool to create the fat libPtex.a archive on OS X.