CMakeLists.txt: do not override user-specified CMAKE_BUILD_TYPE
When building a larger project that has ptex as a submodule, then if one
runs cmake -DCMAKE_BUILD_TYPE=Debug ... at the top level, then the debug
setting for the entire project is overridden by the explicit override of
CMAKE_BUILD_TYPE in ptex's CMakeLists.txt.

Only set CMAKE_BUILD_TYPE when it is unspecified.

Closes #28